add test to catch bug in future
