+++ server/common/utils.py
@@ -11,6 +11,7 @@ from flask import json
 from urllib.parse import urlsplit, urljoin
 import numpy as np
 import pandas as pd
+
 from server.common.errors import ConfigurationError
 
 
@@ -172,3 +173,5 @@ def import_plugins(plugin_module):
         logging.debug(f"No plugins found in module: {plugin_module}: {str(e)}")
 
     return loaded_modules
+
+
